feat: add apartment management functionality with CRUD operations
- Added apartment controller for handling requests related to apartments.
- Implemented apartment service to interact with the apartment repository.
- Created apartment repository for database operations.
- Introduced apartment routes for API endpoints.
- Updated database schema to include default statuses for various tables.
- Added socket.io for real-time communication.
- Enhanced user and resident seeders with additional test data.
- Updated package.json files to include necessary dependencies.

+++ backend/src/models/news.js
@@ -23,7 +23,7 @@ module.exports = (sequelize, DataTypes) => {
         type: DataTypes.TEXT
       },
       category: {
-        type: DataTypes.INTEGER
+        type: DataTypes.TINYINT
       }
     },
     {
